Fix errors in deleting routes and route rules
During the reapply of nmstate templates, the mismatch in routes
and route rules will result in deletion of some routes and route
rules. The deletion of some of the routes conflicts with the
addition of new entries. In order to avoid the conflict, when
there is a difference between the existing route/route rules and
the desired ones, the existing routes and route rules were flushed
and the new entries were made.