migrate: refactor the approach for migration of providers
The earlier approach uses the options like `enable_migration` and
`clean_migration`. With the usage of reboot in the migration process,
these options are no longer needed. The refactored approach is to backup
and remove the network configurations, driver bindings, dpdk map files,
sriov map files configured by the present provider. After the backup and
cleanup of the current provider, os-net-config shall be re-run with the
new provider to complete the migration.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>