migrate: Refactor the approach for migration of providers
The earlier approach uses the options like `enable_migration` and
`clean_migration`. With the usage of reboot in the migration process,
these options are no longer needed. The refactored approach is to backup
and remove the network configurations, driver bindings, dpdk map files,
sriov map files configured by the present provider and then create a
service file which shall start with the new provider. After this
the node is rebooted and when the system comes up the
`os-net-config-migrate` service will configure with the provider.
If the configuration with the new provider fails, then `--rollback`
option allows restoring the configurations from the backed up config files.

Signed-off-by: Karthik Sundaravel <ksundara@redhat.com>